flask-apis-for-model/test-app/requirements.txt
Flask==1.0.3
Werkzeug==0.14.1
requests==2.24.0
Pillow==7.2.0
